separate remote API for Direct Messages, #94
