+++ src/main/java/pityoulish/jrmi/api/RegistryNames.java
@@ -15,10 +15,7 @@ public enum RegistryNames
   MESSAGE_BOARD("RemoteMessageBoard"),
 
   /** The name of the {@link RemoteTicketIssuer} in the registry. */
-  TICKET_ISSUER("RemoteTicketIssuer"),
-
-  /** The name of the {@link RemoteOutletManager} in the registry. */
-  OUTLET_MANAGER("RemoteOutletManager");
+  TICKET_ISSUER("RemoteTicketIssuer");
 
 
   /** The name for looking up or registering this remote object. */
