+++ src/main/java/pityoulish/jrmi/api/dm/DirectMessageOutlet.java
@@ -3,7 +3,7 @@
  * terms of the Creative Commons CC0 1.0 Universal license.
  * https://creativecommons.org/publicdomain/zero/1.0/
  */
-package pityoulish.jrmi.api;
+package pityoulish.jrmi.api.dm;
 
 import java.rmi.Remote;
 import java.rmi.RemoteException;
