+++ src/main/java/pityoulish/jrmi/api/dm/package.html
@@ -1,0 +1,10 @@
+<html>
+<!--
+  This work is released into the Public Domain under the
+  terms of the Creative Commons CC0 1.0 Universal license.
+  https://creativecommons.org/publicdomain/zero/1.0/
+-->
+<body>
+Remote API for the Direct Message extension of the Message Board.
+</body>
+</html>
