+++ src/main/java/pityoulish/jrmi/api/package.html
@@ -5,7 +5,7 @@
   https://creativecommons.org/publicdomain/zero/1.0/
 -->
 <body>
-Remote API for the Message Board and related features.
+Remote API for the Message Board.
 
 <p>
 This package comprises remote interfaces, data object interfaces,
